fix: 调整饼图大小
